# public/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "University of Oslo Library Services Middleware (LSM)",
        "version": "0.2.0"
    },
    "paths": {
        "/alma/search": {
            "get": {
                "tags": [
                    "Alma"
                ],
                "summary": "Search Alma Bib records using SRU.",
                "description": "Search using SRU. Max 10000 records returned. Pagination: If there's no more results, `next` will be null. Otherwise `next` will hold the value to be used with `first` to get the next batch of results.",
                "operationId": "App\\Http\\Controllers\\AlmaController::search",
                "parameters": [
                    {
                        "name": "query",
                        "in": "query",
                        "description": "CQL query string",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "start",
                        "in": "query",
                        "description": "First document to retrieve, starts at 1.",
                        "schema": {
                            "type": "integer",
                            "minimum": 1
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Number of documents to retrieve, defaults to 10, maximum is 50.",
                        "schema": {
                            "type": "integer",
                            "maximum": 50,
                            "minimum": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success"
                    }
                }
            }
        },
        "/alma/records/{id}": {
            "get": {
                "tags": [
                    "Alma"
                ],
                "summary": "Find Alma Bib record by MMS ID.",
                "description": "Get details about a single record",
                "operationId": "App\\Http\\Controllers\\AlmaController::getRecord",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Alma ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "raw",
                        "in": "query",
                        "description": "Set to true to return the raw MARC21 record.",
                        "schema": {
                            "type": "boolean",
                            "default": false
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "An AlmaRecord"
                    }
                }
            }
        },
...
        "/primo/records/{id}": {
            "get": {
                "tags": [
                    "Primo"
                ],
                "summary": "Find Primo record by PNX ID",
                "description": "Get details about a single record",
                "operationId": "App\\Http\\Controllers\\PrimoController::getRecord",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Primo PNX ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "raw",
                        "in": "query",
                        "description": "Set to true to return the raw PNX record.",
                        "schema": {
                            "type": "boolean",
                            "default": false
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A PrimoRecord"
                    }
                }
            }
        },
...
    },
...
}